Moving detach to sparseml/pytorch/utils
